+++ scripts/rebuild_pytest_fixtures.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-PYTEST_REBUILD_FIXTURES=True pytest -s -k yaralyze_with -vv
+PYTEST_REBUILD_FIXTURES=True pytest -k yaralyze_with -vv "$@"